v1.7.1
Refatoração do código existente
