+++ Modulos/data_hora_br.py
@@ -1,6 +1,9 @@
 from datetime import datetime
 
 
+__all__ = ['DatasBr']
+
+
 class DatasBr:
     def __init__(
             self,
