+++ Modulos/janelas/__init__.py
@@ -1,0 +1,5 @@
+from Modulos.janelas.janela_retratil import *
+from Modulos.janelas.janela_sorteios import *
+from Modulos.janelas.paineldelogs import *
+from Modulos.janelas.janeladeimpressao import *
+from Modulos.janelas.janelaregistrodevencedor import *
