+++ validador_de_cadastros.bat
@@ -1,3 +1,2 @@
-@echo off
-call .venv\Scripts\activate 
-py main.py
+C:\Users\Edimar\Documents\GitHub\Validador_de_cadastros\.venv\Scripts\python.exe main.py
+
